requirements.txt
Flask==2.3.3
gunicorn==20.1.0
mysql-connector-python==8.3.0